pyproject.toml
[project]
name = "dqn-stock"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "sqlalchemy",
    "python-dotenv",
    "pymysql",
    "pandas",
    "pandas-stubs==2.3.2.250827",
    "tabulate>=0.9.0",
    "stockstats",
    "matplotlib>=3.10.6",
    "yfinance==0.2.65",
    "alpaca-py>=0.29.2",
    "mplfinance>=0.12.10b0",
]
